+++ Runtime/GameFeedBacks/GameFeedback.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+
+namespace Etienne.Feedback
+{
+    [System.Serializable]
+    public abstract class GameFeedback
+    {
+        public abstract Task Execute(UnityEngine.GameObject gameObject);
+    }
+}