+++ Runtime/GameFeedBacks/PlaySound.cs
@@ -1,0 +1,25 @@
+using System.Threading.Tasks;
+using UnityEngine;
+
+namespace Etienne.Feedback
+{
+    [System.Serializable]
+    [GameFeedback(0, 153, 255, "Audio/PlaySound")]
+    public class PlaySound : GameFeedback
+    {
+
+        [SerializeField] private bool _IsAttached;
+        [SerializeField] private Cue _Cue = new Cue(null);
+
+        public override async Task Execute(GameObject gameObject)
+        {
+            _Cue.Play(_IsAttached ? gameObject.transform : null);
+            await Task.CompletedTask;
+        }
+
+        public override string ToString()
+        {
+            return $"Play {_Cue}{(_IsAttached ? ", attached to transform" : "")}";
+        }
+    }
+}